+++ tests/homeassistant.test.mjs
@@ -1,0 +1,10 @@
+import { createDeviceIdentifierString } from '../app/homeassistant.mjs'
+
+test('createDeviceIdentifierString', () => {
+    const deviceInformation = {
+        'modelType': 'Pingvin',
+        'fanType': 'EC',
+    }
+
+    expect(createDeviceIdentifierString(deviceInformation)).toEqual('enervent-pingvin-ec')
+})